refactor: extract components from App.tsx and improve settings button styling
