+++ example-dimo-auth/src/components/AdvancedTransactionButton/index.ts
@@ -1,0 +1,1 @@
+export * from './AdvancedTransactionButton';
