+++ example-dimo-auth/src/components/UserData/index.ts
@@ -1,0 +1,1 @@
+export * from './UserData';
