+++ example-dimo-auth/src/components/index.ts
@@ -1,2 +1,5 @@
 export * from './Input';
 export * from './ConfigForm';
+export * from './Examples';
+export * from './AdvancedTransactionButton';
+export * from './UserData';
